+++ src/InvestorsExchangeData/Deep.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace Digitonic\IexCloudSdk\InvestorsExchangeData;
+
+use Digitonic\IexCloudSdk\Contracts\IEXCloud;
+use Digitonic\IexCloudSdk\Exceptions\WrongData;
+use Digitonic\IexCloudSdk\Requests\BaseGet;
+
+class Deep extends BaseGet
+{
+    const ENDPOINT = 'deep?symbols={symbol}';
+
+    /**
+     * Create constructor.
+     *
+     * @param  IEXCloud  $api
+     */
+    public function __construct(IEXCloud $api)
+    {
+        parent::__construct($api);
+    }
+
+    /**
+     * @return string
+     */
+    protected function getFullEndpoint(): string
+    {
+        return str_replace('{symbol}', $this->symbol, self::ENDPOINT);
+    }
+
+    /**
+     * @return bool|void
+     */
+    protected function validateParams(): void
+    {
+        if (empty($this->symbol)) {
+            throw WrongData::invalidValuesProvided('Please provide symbol to query!');
+        }
+    }
+}