+++ CHANGELOG.md
@@ -2,6 +2,52 @@
 
 All notable changes to `iex-cloud-sdk` will be documented in this file
 
-## 1.0.0 - 201X-XX-XX
+## v0.1.0-beta.3 - 2019-10-01
 
-- initial release
+### Added
+
+##### Alternative Data
+
+- Crypto
+- CEO Compensation
+
+##### Investors Exchange Data
+
+- TOPS
+- TOPS Last
+- DEEP
+- DEEP Auction
+- DEEP Book
+- DEEP Security Event
+
+## v0.1.0-beta.2 - 2019-09-30
+
+### Added
+
+##### Data Apis
+
+- Data Points
+- Time Series Inventory
+- Time Series Query
+
+## v0.1.0-beta.1 - 2019-09-26
+
+### Added
+
+##### Account
+
+- Metadata
+- Usage
+
+##### API System Metadata
+
+- Status
+
+##### Forex / Currencies
+
+- Exchange Rates
+
+##### Reference Data
+
+- Symbols
+- IEX Symbols